Made progress with picking
Idea is that I track whether a user is trying to pick a face. If they are, the cube is rendered to the back buffer but no transformations are applied. Then the color under the cursor will be the color of the cube's initial configuration which easily gives which face to rotate